test: Add CustomerServiceTest (#9)
Add `CustomerServiceTest` which includes tests for creating customers and retrieving customers. These tests will assert the contents of the returned customer object. It might unnecessary to have both service and controller tests that check the same thing. However, the purpose here is to give examples of what a controller and service test could look like.