fix: Disable bootJar and bootRun in persistence build.gradle.kts
The persistence subproject is not supposed to be an independently executable subproject. This subproject will be used by the api subproject which is runnable. To ensure that we can use gradlew build locally, we must disable gradle tasks bootJar and bootRun because this subproject has the Spring Boot plugin but should be runnable.